+++ apps/web/app/(navigation)/(singleCard)/MainPageTranslated.tsx
@@ -1,6 +1,6 @@
 "use client";
 
-import { TypoStyled, LinkStyled } from "app/(navigation)/HomePageComponents";
+import { TypoStyled } from "app/(navigation)/HomePageComponents";
 import { useTranslate } from "lib/hooks";
 import { signIn, signOut, useSession } from "next-auth/react";
 import { Button } from "ui";
@@ -14,14 +14,10 @@ const ContentWrapperStyled = styled.div`
 `;
 export const MainPageTranslated = () => {
   const { dict, t } = useTranslate("MainPage");
-  const { data: session } = useSession();
 
   return (
     <ContentWrapperStyled>
       <TypoStyled>{t(dict.welcomeText)}</TypoStyled>
-      {!session && (
-        <LinkStyled href="/sign-up">{t(dict.createAccountLink)}</LinkStyled>
-      )}
     </ContentWrapperStyled>
   );
 };
