+++ apps/web/app/(navigation)/HomePageComponents.tsx
@@ -1,6 +1,6 @@
 "use client";
 import { device } from "lib/media-queries";
-import styled from "styled-components";
+import styled, { css } from "styled-components";
 import Link from "next/link";
 import { Card } from "ui";
 
@@ -22,18 +22,6 @@ export const CardStyled = styled(Card)`
   padding: 14px 10px;
   ${device.tablet} {
     padding: 32px 48px;
-  }
-`;
-
-export const LinkStyled = styled(Link)`
-  font-family: "Signika", sans-serif;
-  font-size: 1.3em;
-  font-weight: 600;
-  color: ${({ theme }) => theme.nav.main};
-  margin-top: 35px;
-  padding: 10px;
-  ${device.desktop} {
-    font-size: 2.3em;
   }
 `;
