+++ apps/web/app/(navigation)/budgets/[id]/BudgetDetails/BudgetDetails.styled.tsx
