+++ apps/web/app/(navigation)/budgets/[id]/BudgetDetails/BudgetDetails.tsx
