+++ apps/web/app/(navigation)/budgets/[id]/BudgetDetails/TrendChart.tsx
